feat(web): add filter issues by teams option
